% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AKMap_areas_function.R
\name{AK_MercatorL_fxn}
\alias{AK_MercatorL_fxn}
\title{Alaska Map Functions (Mercator, labeled)}
\usage{
AK_MercatorL_fxn(Path = "internal", ShapeFileName = "internal", label.size)
}
\description{
These function creates base maps of Alaska inclusive of NMFS Reporting Areas, showing Canadian and Russian shorelines.
This function utilizes a shape file that is contained in the JAC tools package (default) that was originally generated by J.Gasper (AKRO, April 2021).
Alternative shape files can be used by specifing a path and file name to the boundary shape file (.shp).  \cr
}
\details{
Note that the warning message:\cr
\emph{Warning message: In st_centroid.sf(nmfs_areas_shapefile): st_centroid assumes attributes are constant over geometries of x} \cr
Indicates that the lables are not located exactly on the centroid for the NMFS Reporting Area (in addition some were intentially moved). \cr

Additionally note the message \emph{"although coordinates are longitude/latitude, st_intersection assumes that they are planar" }
indicates that spatial coordinates are on a flat space ({\eqn{R^2}}, not an ellipsoid or sphere on {\eqn{S^2}}). Package s2 is not in use. \cr
See also sf package: \url{https://cran.r-project.org/web/packages/sf/index.html} and s2 package \url{https://cran.r-project.org/web/packages/s2/index.html} \cr

There are 5 maps contained in the JACtools package: \itemize{\item 2 Mercator with NMFS Reporting Areas (with and without labels: AK_MercatorL_fxn, AK_Mercator_fxn)
\item 2 Robinson with NMFS Reporting Areas (with and without labels: AK_RobinsonL_fxn, AK_Robinson_fxn)
\item 1 Mercator without NMFS Reporting Areas (AKMap_fxn)}
\url{https://www.icsm.gov.au/education/fundamentals-mapping/projections/commonly-used-map-projections}
}
\examples{
AK_MercatorL_fxn()
}
\keyword{AK}
\keyword{NMFS_Area}
\keyword{map}
