% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AK_onlyMap_function.R
\name{AK_Robinson_fxn}
\alias{AK_Robinson_fxn}
\title{An Alaska Map Function (Robinson, unlabeled)}
\usage{
AK_Robinson_fxn(Path = "internal", ShapeFileName = "internal")
}
\description{
This function creates a base map of Alaska inclusive of NMFS Reporting Areas, showing Canadian and Russian shorelines.
  This function utilizes a shape file that is contained in the JAC tools package (default) that was originally generated by J.Gasper (AKRO, April 2021). 
  Alternative shape files can be used by specifing a path and file name to the boundary shape file (.shp).  
Note that the warning message:
    Warning message: In st_centroid.sf(nmfs_areas_shapefile): st_centroid assumes attributes are constant over geometries of x
 Indicates that the lables are not located exactly on the centroid for the NMFS Reporting Area (in addition some were intentially moved).
}
\details{
There are 5 maps contained in the JACtools package: 2 Mercator with NMFS Reporting Areas (with and without labels: AK_MercatorL_fxn, AK_Mercator_fxn)
                                                    2 Robinson with NMFS Reporting Areas (with and without labels: AK_RobinsonL_fxn, AK_Robinson_fxn)
                                                    1 Mercator without NMFS Reporting Areas (AKMap_fxn)
(https://www.icsm.gov.au/education/fundamentals-mapping/projections/commonly-used-map-projections)
}
\examples{
AK_Robinson_fxn()
}
\keyword{AK}
\keyword{NMFS_Area}
\keyword{map}
