% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alaska_shapefile.R
\docType{data}
\name{alaska_shapefile}
\alias{alaska_shapefile}
\title{A base map of Alaska shoreline including Canada and Russia}
\format{
A shapefile (.shp)
}
\usage{
data(alaska)
}
\description{
This file contains the Alaska shoreline shape file created by J.Gasper (April 2020).
This file can be used to generate a ggplot object showing Alaska.
See AK_onlyMap_fxn()
}
\examples{
JACtools::AK_onlyMap_fxn()

}
\keyword{alaska}
\keyword{shapefile}
