% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode_Functions.R
\name{anti_Mode_alpha}
\alias{anti_Mode_alpha}
\title{Compute the NOT Mode of character data}
\usage{
anti_Mode_alpha(x)
}
\arguments{
\item{x}{A vector cinsiting of two character values for which the Mode(x) is computed. x must be a vector of character values, missing values are dropped from computations; for numeric data use Mode.numeric().}
}
\value{
a character value, the value that is NOT the mode of x
}
\description{
This function returns the anti-mode of a vector(x).
}
\examples{
x <- c("a","a","b","a","a","b")
anti_Mode_alpha(x)

}
\keyword{Mode}
