% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode_Functions.R
\name{anti_Mode_numeric}
\alias{anti_Mode_numeric}
\title{Compute the NOT Mode of numeric data}
\usage{
anti_Mode_numeric(x)
}
\arguments{
\item{x}{A vector consisting of two values for which the Mode(x) is computed. x must be a vector of numeric values, missing values are dropped from computations; for character data use Mode.alpha().}
}
\value{
a numeric value, the value that is not the mode of x
}
\description{
This function returns the value that is NOT the mode (anti-mode) of a numeric vector(x).
Note this is only valid for vectors with two possible values
}
\examples{
x <- c(5,4,5,4,5,5,4,4,5,5,4,5,4,4,5,4,5,5,5,4,4,5,5,4,5)
anti_Mode_numeric(x)

}
\keyword{anti-Mode}
