% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_plots_function.R
\name{arrange_plots}
\alias{arrange_plots}
\title{Arrange Multiple Plots on a Page}
\usage{
arrange_plots(..., nrow = NULL, ncol = NULL, as.table = FALSE)
}
\arguments{
\item{nrow}{Number of rows of plots, defaults to NULL

depends: ggplot2,
grid}

\item{ncol}{Number of columns of plots, defaults to NULL}

\item{plotlist}{A list of plots to be arranged. Defaults to NULL.}
}
\description{
This function places a list of ggplot objects into a single object.
taken from RBloggers, 3July12, combining ggplot images \url{https://www.r-bloggers.com/combining-ggplot-images/}
}
\examples{
x <- rnorm(100, 10, 3)
y <- rnorm(200, 10, 3)
test.dat <- as.data.frame(cbind(x, y))
fig.1 <- ggplot2::ggplot(test.dat, aes(x,y)) + geom_point()
fig.2 <- ggplot2::ggplot(test.dat[which(test.dat$x > 5),], aes(x,y)) + geom_point()

  p1 <- qplot(wt, mpg, data=mtcars)
  p2 <- ggplot(diamonds, aes(price, colour = cut)) +
                     geom_density()
   arrange_plots(p1,p2)

}
\keyword{ggplot2}
\keyword{multiplot}
