% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_bins_Functions.R
\name{expand_binned}
\alias{expand_binned}
\title{Expand binned frequency data to one record per item}
\usage{
expand_binned(x)
}
\arguments{
\item{x}{A dataframe with one column called 'FREQUENCY'. Each row of the dataframe is copied 'FREQUENCY' times.}
}
\value{
a dataframe with sum(FREQUENCY) number of records
}
\description{
This function takes binned data and expands it to one record per item, returning an expanded data frame.
For large data sets, this may take a while
}
\examples{
Covariate <- c('Red', 'Blue','Red','Red','Blue','Blue')
FREQUENCY <-  c(5,4,5,3,4,1)
ID <- c(1,2,3,4,5,6)

test.dat <- data.frame(ID,Covariate,FREQUENCY)
for(i in 1:length(test.dat$ID)) { 
 if(i==1) expanded.test.dat <- expand_binned(test.dat)
 if (i>1) expanded.test.dat <- rbind(expanded.test.dat, expand_binned(test.dat))
}
test.dat
expanded.test.dat
}
\keyword{binned,}
\keyword{frequency,}
\keyword{length}
