% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_function.R
\name{get_connected}
\alias{get_connected}
\title{Database Connect Function}
\usage{
get_connected(schema = "AFSC")
}
\arguments{
\item{schema}{The name of the schema to establish a connection with, in quotes. IMPORTANT: TNS.names file must include target schema. Defaults to AFSC.}
}
\description{
Opens connection to Oracle database schemas (RODBC)
This function opens RODBC connections to database schemas such as the AFSC schemas. Users will be prompted for username and password.
schema defaults to 'AFSC'
}
\examples{
get_connected()
}
\keyword{connection}
\keyword{database}
