% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Return_n_function.R
\name{give_n}
\alias{give_n}
\title{Place Sample Size on Boxplots at Mean y-value}
\usage{
give_n(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function returns the length (number of observations, n) and the mean of a vector. 
When used with ggplot, it can to place the number of observations within the box of a box-and-whisker plot into the box at the mean value of the observations.
}
\examples{
color.dat <- c(rep('A', 10),rep('B', 10))
y.dat <- c(rnorm(10,2, 2), rnorm(10,5,5))
example.dat <- data.frame(color.dat,y.dat)
(ggplot2::ggplot(example.dat, aes(y=y.dat, x=color.dat, fill=color.dat))
		+ geom_boxplot(color='grey30', aes(fill=color.dat))
		+ stat_summary(fun.data = give_n, geom = "text", position = position_dodge(width = 0.75), size = 5, color='red')
		+ scale_fill_manual('', values=c("grey80","grey60"))
		+ theme_bw())
}
\keyword{boxplot}
\keyword{sample}
\keyword{size;}
