% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_legend_function.R
\name{grid_arrange_shared_legend}
\alias{grid_arrange_shared_legend}
\title{Share a Common Legend for Multiple Plots on a Page}
\usage{
grid_arrange_shared_legend(...)
}
\arguments{
\item{plotlist}{A list of plots to be combined. Defaults to NULL.}
}
\description{
This function places a list of ggplot objects into a single object and
   provides a shared legend for those plots.
   Plots are stacked - one column.
}
\examples{

x <- rnorm(200, 10, 3)
y <- rnorm(200, 10, 3)
z <- rep(c('Up', 'Down'),200)
test.dat <- as.data.frame(cbind(x, y,z))
fig.1 <- ggplot2::ggplot(test.dat, aes(x,y)) + geom_point(aes(color=z))
fig.2 <- ggplot2::ggplot(test.dat[which(test.dat$x > 5),], aes(x,y)) + geom_point(aes(color=z))
grid_arrange_shared_legend(fig.1, fig.1)
}
\keyword{ggplot2}
\keyword{legend}
\keyword{multiplot}
