% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haversine_distance_function.R
\name{hav_dist}
\alias{hav_dist}
\title{A Haversine Distance Function}
\usage{
hav_dist(long1, lat1, long2, lat2, radius = 6371)
}
\arguments{
\item{long1}{longitude for first point}

\item{lat1}{latitude for first point}

\item{long2}{longitude for second point}

\item{lat2}{latitude for first point}

\item{radius}{the radius of the earth; units of computed distance will be the same as the units of R. Dafaults to 6371 kilometers.}
}
\description{
This function computes the great circle distance between two points (long, lat). 
Computations default to kilometers (for the WGS84 ellipsoid). This code was 
  adapted from https://en.wikipedia.org/wiki/Great-circle_distance;  
  https://en.wikipedia.org/wiki/Haversine_formula;  http://www.movable-type.co.uk/scripts/latlong.html  
In addition, see in the geosphere package  distm(p1, p2, fun = distVincentyEllipsoid) and distHaversine()
}
\examples{
hav_dist(0,0,45,45)
distHaversine(c(0,0), c(45,45))/1000
}
\keyword{distance,}
\keyword{haversign}
