% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heart_function.R
\name{heart_fxn}
\alias{heart_fxn}
\title{A Heart Function}
\usage{
heart_fxn(name = NULL, message = "Happy Valentine's Day", text.color = "blue")
}
\arguments{
\item{name}{A name to be included under the message. Defaults to NULL}

\item{message}{A message to write on the heart. Defaults to 'Happy Valentine's Day'}

\item{text.color}{Text color. Defaults to 'blue'}
}
\description{
This function places a message on a heart-shaped plot
I have no idea from whence this function came to me.
}
\examples{
heart_fxn('Jen', 'Hello', 'green')
heart_fxn('Hello')
heart_fxn(message = 'Hello')
}
\keyword{heart}
