% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_redef_function.R
\name{mean_2sd}
\alias{mean_2sd}
\title{A Function to set the hinges of boxplt equal to 1.96 Std. Dev}
\usage{
mean_2sd(x)
}
\arguments{
\item{x}{a vector of numeric data}
}
\description{
This function re-defines the hinges of a boxplot to equal the mean +/- 1.96 standard deviation of the data
and the whiskers to be the range of the data (min and max). \cr
If datapoints are normally distributed, this would equate to +/- 95\% Confidence Interval \cr
See also mean.sd for box defined by 1 std. dev
}
\examples{
a <- rnorm(50, 0, 1)
b <- rnorm(120, 0.5, 1.5)
c <- rnorm(70, 1, 0.5)
d <- rnorm(15, 0,1)
group <- factor(rep(1:4, c(50, 120, 70, 15)))

mydata <- data.frame(c(a,b,c,d), group)
names(mydata) <- c("value", "group")

p2 <- ggplot(aes(y = value, x = factor(group)), data = mydata)
p2 <- (p1 + stat_summary(fun.data = mean.2sd, geom = "boxplot") 
		+ geom_jitter(position=position_jitter(width=.2), size=1.5) 
		+ ggtitle("Boxplot con media, 2Std. Dev. (95\%CI), valore min. e max.") + xlab("Gruppi") + ylab("Valori"))
p2

}
\keyword{Std.Dev,}
\keyword{Std.Err}
\keyword{boxplot,}
