% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode_Functions.R
\name{Mode_alpha}
\alias{Mode_alpha}
\title{Compute the Mode of character data}
\usage{
Mode_alpha(x)
}
\arguments{
\item{x}{A vector of character values for which the Mode(x) is computed. x must be a vector of character values, missing values are dropped from computations; for numeric data use Mode.numeric().}
}
\value{
a character value, the mode of x
}
\description{
This function returns the mode of a vector(x).
}
\examples{
x <- c("a","a","b","c","a","c")
Mode_alpha(x)

}
\keyword{Mode}
