% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode_Functions.R
\name{Mode_numeric}
\alias{Mode_numeric}
\title{Compute the Mode of numeric data}
\usage{
Mode_numeric(x)
}
\arguments{
\item{x}{A vector of values for which the Mode(x) is computed. x must be a vector of numeric values, missing values are dropped from computations; for character data use Mode.alpha().}
}
\value{
a numeric value, the mode of x
}
\description{
This function returns the mode of a numeric vector(x).
}
\examples{
x <- c(5,4,7,3,5,8,4,1,3,5,7,2,6,4,6,4,5,5,5,7,4,5,8,6,5)
Mode_numeric(x)

}
\keyword{Mode}
