% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot_function.R
\name{multiplot}
\alias{multiplot}
\title{Put Multiple Plots on a Page}
\usage{
multiplot(..., plotlist = NULL, cols)
}
\arguments{
\item{plotlist}{A list of plots to be combined. Defaults to NULL.}

\item{cols}{Number of columns of plots.}
}
\description{
This function places a list of ggplot objects into a single object. 
taken from RBloggers, 3july12, combining ggplot images, https://www.r-bloggers.com/2012/07/combining-ggplot-images/
}
\examples{

x <- rnorm(100, 10, 3)
y <- rnorm(200, 10, 3)
test.dat <- as.data.frame(cbind(x, y))
fig.1 <- ggplot2::ggplot(test.dat, aes(x,y)) + geom_point()
fig.2 <- ggplot2::ggplot(test.dat[which(test.dat$x > 5),], aes(x,y)) + geom_point()
multiplot(fig.1, fig.1, cols=2)
}
\keyword{ggplot2}
\keyword{multiplot}
