% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfs_areas_shapefile.R
\docType{data}
\name{nmfs_areas_shapefile}
\alias{nmfs_areas_shapefile}
\title{A base map of Alaska NMFS Reporting Areas}
\format{
A shapefile (.shp)
}
\usage{
data(areas)
}
\description{
This file contains the Alaska NMFS Reporting Area boundaries shape file.
NMFS Reporting Area boundaries are based on a shape file created by J.Gasper (April 2020).
This file can be used to generate a ggplot object showing AK NMFS Reporting Areas
See AK_AreasMap_fxn()
}
\examples{

(JACtools::AK_AreasMap_fxn()

}
\keyword{shapefile}
