% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outersect_function.R
\name{outersect.fxn}
\alias{outersect.fxn}
\title{Find the anti-intersection of two vectors}
\usage{
outersect.fxn(x, y, ...)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}
}
\description{
This function returns the values that are not in both vectors. \cr
}
\examples{
color.dat <- c(rep('A', 5),rep('B', 5))
y.dat <- c(rnorm(5,2, 2), rnorm(5,5,5))
x.dat <- c(rnorm(5,20, 2), rnorm(5,50,5))
df1 <- data.frame(color.dat,x.dat)
df2 <- data.frame(color.dat,y.dat)
outersect(names(df1),names(df2))
}
\keyword{anti}
\keyword{intersection}
