% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_legend_function.R
\name{shift_legend}
\alias{shift_legend}
\title{Shift the Legend of a Facetted Plot to an Empty Facet}
\usage{
shift_legend(p)
}
\arguments{
\item{p}{A ggplot object.}
}
\description{
This function places the legend of a facetted plot with an uneven number of facets
into the empty space (the empty facet area). \cr
This function is taken from stack overflow: \url{https://stackoverflow.com/questions/54438495/shift-legend-into-empty-facets-of-a-faceted-plot-in-ggplot2} \cr
See also \url{https://stackoverflow.com/questions/28496764/add-textbox-to-facet-wrapped-layout-in-ggplot2}  \cr
and the cowplot package \url{https://cran.r-project.org/web/packages/cowplot/vignettes/introduction.html}
}
\examples{

p <- (ggplot2::ggplot(diamonds, aes(x = carat, fill = cut)) 
      + geom_density(position = "stack") + facet_wrap(~ color) )
library(grid)
grid.draw(shift_legend(p))

p.new <- p +
  guides(fill = guide_legend(title.position = "top",
                            label.position = "bottom",
                            nrow = 1)) +
  theme(legend.direction = "horizontal")
grid.draw(shift_legend(p.new))
}
\keyword{cowplot}
\keyword{facet}
\keyword{ggplot2}
\keyword{legend}
