% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substrRight_function.R
\name{substrRight}
\alias{substrRight}
\title{A Right-handed Substring Function}
\usage{
substrRight(x, n, k)
}
\arguments{
\item{x}{a vector, if not a character vector, will be coerced using as.character}

\item{n}{the number of elements from the right margin to start counting from}

\item{k}{the number of elements to include in the resulting string}
}
\description{
This function returns a portion of a string, counted from the right-hand margin.
Useful for strings of uneven length where the trailing digits are of interest.
This function is similar to base::substr() except that the offset if counted from the right.
}
\examples{
substr("abcdef", 2, 4)
substrRight(X, 3, 2)
X <- c(3846250)
substrRight(X, 4, 2)
}
\keyword{substr}
